+++ inst/htmlwidgets/amChart4.js
@@ -121771,6 +121771,7 @@ class AmHorizontalBarChart extends React.PureComponent {
         threeD = this.props.threeD,
         chartLegend = this.props.legend,
         category = this.props.category,
+        categories = this.props.data[category],
         values = this.props.values,
         data = HTMLWidgets.dataframeToD3(this.props.data),
         dataCopy = HTMLWidgets.dataframeToD3(_utils__WEBPACK_IMPORTED_MODULE_13__["subset"](this.props.data, [category].concat(values))),
@@ -121925,6 +121926,67 @@ class AmHorizontalBarChart extends React.PureComponent {
           Shiny.setInputValue(shinyId + ":rAmCharts4.dataframe", chart.data);
           Shiny.setInputValue(shinyId + "_change", null);
         }
+      });
+    }
+    /* ~~~~\  Shiny message handler for horizontal bar chart  /~~~~ */
+
+
+    if (window.Shiny) {
+      Shiny.addCustomMessageHandler(shinyId + "bar", function (newdata) {
+        var tail = " is missing in the data you supplied!"; // check that the received data has the 'category' column
+
+        if (!newdata.hasOwnProperty(category)) {
+          console.warn("updateAmBarChart: column \"".concat(category, "\"") + tail);
+          return null;
+        } // check that the received data has the necessary categories
+
+
+        var ok = true,
+            i = 0;
+
+        while (ok && i < categories.length) {
+          ok = newdata[category].indexOf(categories[i]) > -1;
+
+          if (!ok) {
+            console.warn("updateAmBarChart: category \"".concat(categories[i], "\"") + tail);
+          }
+
+          i++;
+        }
+
+        if (!ok) {
+          return null;
+        } // check that the received data has the necessary 'values' columns
+
+
+        i = 0;
+
+        while (ok && i < values.length) {
+          ok = newdata.hasOwnProperty(values[i]);
+
+          if (!ok) {
+            console.warn("updateAmBarChart: column \"".concat(values[i], "\"") + tail);
+          }
+
+          i++;
+        }
+
+        if (!ok) {
+          return null;
+        } // update chart data
+
+
+        var tnewdata = HTMLWidgets.dataframeToD3(newdata);
+
+        for (var r = 0; r < data.length; ++r) {
+          for (var v = 0; v < values.length; ++v) {
+            chart.data[r][values[v]] = tnewdata[r][values[v]];
+          }
+        }
+
+        chart.invalidateRawData();
+        Shiny.setInputValue(shinyId + ":rAmCharts4.dataframe", tnewdata);
+        Shiny.setInputValue(shinyId + "_change", null);
       });
     }
     /* ~~~~\  category axis  /~~~~ */
@@ -124544,6 +124606,7 @@ class AmRadialBarChart extends React.PureComponent {
     var theme = this.props.theme,
         chartLegend = this.props.legend,
         category = this.props.category,
+        categories = this.props.data[category],
         values = this.props.values,
         minValue = this.props.minValue,
         maxValue = this.props.maxValue,
@@ -124698,6 +124761,67 @@ class AmRadialBarChart extends React.PureComponent {
           Shiny.setInputValue(shinyId + ":rAmCharts4.dataframe", chart.data);
           Shiny.setInputValue(shinyId + "_change", null);
         }
+      });
+    }
+    /* ~~~~\  Shiny message handler for radial bar chart  /~~~~ */
+
+
+    if (window.Shiny) {
+      Shiny.addCustomMessageHandler(shinyId + "bar", function (newdata) {
+        var tail = " is missing in the data you supplied!"; // check that the received data has the 'category' column
+
+        if (!newdata.hasOwnProperty(category)) {
+          console.warn("updateAmBarChart: column \"".concat(category, "\"") + tail);
+          return null;
+        } // check that the received data has the necessary categories
+
+
+        var ok = true,
+            i = 0;
+
+        while (ok && i < categories.length) {
+          ok = newdata[category].indexOf(categories[i]) > -1;
+
+          if (!ok) {
+            console.warn("updateAmBarChart: category \"".concat(categories[i], "\"") + tail);
+          }
+
+          i++;
+        }
+
+        if (!ok) {
+          return null;
+        } // check that the received data has the necessary 'values' columns
+
+
+        i = 0;
+
+        while (ok && i < values.length) {
+          ok = newdata.hasOwnProperty(values[i]);
+
+          if (!ok) {
+            console.warn("updateAmBarChart: column \"".concat(values[i], "\"") + tail);
+          }
+
+          i++;
+        }
+
+        if (!ok) {
+          return null;
+        } // update chart data
+
+
+        var tnewdata = HTMLWidgets.dataframeToD3(newdata);
+
+        for (var r = 0; r < data.length; ++r) {
+          for (var v = 0; v < values.length; ++v) {
+            chart.data[r][values[v]] = tnewdata[r][values[v]];
+          }
+        }
+
+        chart.invalidateRawData();
+        Shiny.setInputValue(shinyId + ":rAmCharts4.dataframe", tnewdata);
+        Shiny.setInputValue(shinyId + "_change", null);
       });
     }
     /* ~~~~\  category axis  /~~~~ */